A little more work on the file scanner
Added a function that prevents the bot from scanning his own messages.
Updated the error message for not being able to load a cog in the Miscellaneous cog.
Added a system that downloads message attachments to a directory in the downloads folder named with
    the ID of the message, preventing overriding of existing files with the same name.
Added a system to the notifying message that gives the original text of the message, if any.

Now to just build the file scanner itself.