Updated moderation system
Fixed an issue where if you typed out a long word in a message to specifically abuse BOB's usage of tokenization in his scanner. Before, if you blocked the word "among", and a user typed out the word "dddamongddd", BOB couldn't catch it. Now he checks if the word is plainly in the token, and the word is not whitelisted. This will, of course, lead to more false positives without a proper whitelist.